// src/app/modules/issue/issue.interface.ts
export interface IIssue {
  id?: number;
  title: string;
  description: string;
  type: 'bug' | 'feature_request';
  status?: 'open' | 'in_progress' | 'resolved';
  reporter_id?: number;
  created_at?: Date;
  updated_at?: Date;
}

export interface IIssueUpdate {
  title?: string;
  description?: string;
  type?: 'bug' | 'feature_request';
  status?: 'open' | 'in_progress' | 'resolved';
}